feat: add LearnWithDoppler component
