+++ src/components/Dashboard/LearnWithDoppler/Carousel/Slide/Slide.js
