+++ src/components/Dashboard/LearnWithDoppler/Carousel/Slide/Slide.test.js
